requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
